Refactor unittest test classes into own files
