+++ tox.ini
@@ -13,7 +13,7 @@ commands =
     coverage erase
 
     # Unit tests
-    {envpython} -m pytest --cov osmviz --cov test --cov-append test/unit/test_manager.py {posargs}
+    {envpython} -m pytest --cov osmviz --cov test --cov-append test/unit {posargs}
     {envpython} -m pytest --cov osmviz --cov test --cov-append test/functional/test_manager.py {posargs}
     coverage xml
